Add decrease count functionality to cart
Summary: Cart now has buttons to remove items, reducing the count of the line item and finally destroying the line completely when the count == 0.  When an item's count is reduced the row effect will highlight red in the cart and fade back to green.  When a cart has 0 items in will slide into hidden.

Test Plan: N/A

Reviewers: aaron

Reviewed By: aaron

Differential Revision: http://phabricator.sazboom.com/D15